requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
awesome-slugify==1.6.5
beautifulsoup4==4.9.3
bleach==4.1.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
cycler==0.10.0
dataframe-image==0.1.1
defusedxml==0.7.1
entrypoints==0.3
idna==3.2
ipython-genutils==0.2.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==7.0.1
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.2
loguru==0.5.3
MarkupSafe==2.0.1
matplotlib==3.4.3
mistune==0.8.4
multidict==5.1.0
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
numpy==1.21.2
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
Pillow==8.3.1
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
pyzmq==22.2.1
regex==2021.8.28
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
testpath==0.5.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
Unidecode==0.4.21
urllib3==1.26.6
webencodings==0.5.1
yarl==1.6.3
